+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * added `getTypedExtensions`, `getTypes`, and `getTypeGuessers` to `Symfony\Component\Form\Test\FormIntegrationTestCase`
  * added `FormPass`
  * deprecated `choice_attr` option as array of nested arrays mapped by indexes
+ * deprecated `choice_attr` option as string or `PropertyPath` instance
 
 3.2.0
 -----